Fix parameters with - and remove content to shortcode
As today the parameter with `__` are not revert back as `-` in the preview.
Also I saw that the content parameter of the shortcode is sent to the rendering but is unused as parameter and is something private. Also passing content generate hundreds of shortcode parameter as it parses the HTML and create a bit of mess.
There is any hope for a fast release with those fix?